refactor(dashboard): rename folder
